Added the kappa term, to get a bigger imbalance response at the material interface, moved some field around to be consistent

+++ pkd.h
@@ -1040,7 +1040,7 @@ static inline float pkdSoft( PKD pkd, PARTICLE *p ) {
 static inline FIO_SPECIES pkdSpecies( PKD pkd, PARTICLE *p ) {
     return pkd->particles.species(p);
 }
-static inline float pkdiMat( PKD pkd, PARTICLE *p ) {
+static inline int pkdiMat( PKD pkd, PARTICLE *p ) {
     return pkd->particles.iMat(p);
 }
